move script caching code to xonsh/tools.py
